requirements.txt
streamlit
python-dotenv
requests
langchain
openai
sqlite3
transformers
accelerate
sentencepiece